requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
conda==4.9.2
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
file-read-backwards==2.0.0
filelock==3.0.12
future==0.18.2
idna==2.9
imageio==2.9.0
imageio-ffmpeg==0.4.2
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
numpy-stl==2.12.0
olefile==0.46
opencv-python==4.4.0.46
packaging==20.4
pandas==1.2.3
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.15
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.23.0
ruamel-yaml==0.15.87
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.1
sip==4.19.13
six==1.14.0
sklearn==0.0
stl==0.0.3
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
urllib3==1.25.8
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
